pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "aex"
authors= [{name = "Remi Louf", email = "remi@thetypicalset.com"}]
description = "Sampling with Blackjax on Aesara"
requires-python = ">=3.7"
keywords=[
    "probability",
    "machine learning",
    "statistics",
    "mcmc",
    "sampling",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: MacOS",
    "Operating System :: POSIX",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Education",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Mathematics",
]
dependencies = [
    "aesara",
    "aeppl",
    "jax",
    "jaxlib",
    "blackjax-nightly"
]
version="0.0.2"

[project.urls]
homepage = "https://github.com/rlouf/aex"
repository = "https://github.com/rlouf/aex"

[project.readme]
file="README.md"
content-type = "text/markdown"
